+++ gettext-tools/src/format-c.c
@@ -33,16 +33,16 @@
 #include "format-invalid.h"
 
 #define INVALID_C99_MACRO(directive_number) \
-  xasprintf (_("In the directive number %u, the token after '<' is not the name of a format specifier macro. The valid macro names are listed in ISO C 99 section 7.8.1."), directive_number)
+  xasprintf (_("In the directive number %zu, the token after '<' is not the name of a format specifier macro. The valid macro names are listed in ISO C 99 section 7.8.1."), directive_number)
 
 #define INVALID_ANGLE_BRACKET(directive_number) \
-  xasprintf (_("In the directive number %u, the token after '<' is not followed by '>'."), directive_number)
+  xasprintf (_("In the directive number %zu, the token after '<' is not followed by '>'."), directive_number)
 
 #define INVALID_SIZE_SPECIFIER(directive_number) \
-  xasprintf (_("In the directive number %u, the argument size specifier is invalid."), directive_number)
+  xasprintf (_("In the directive number %zu, the argument size specifier is invalid."), directive_number)
 
 #define INVALID_IGNORED_ARGUMENT(referenced_arg, ignored_arg) \
-  xasprintf (_("The string refers to argument number %u but ignores argument number %u."), referenced_arg, ignored_arg)
+  xasprintf (_("The string refers to argument number %zu but ignores argument number %zu."), referenced_arg, ignored_arg)
 
 /* Execute statement if memory allocation function returned NULL.  */
 #define IF_OOM(allocated_ptr, statement)  /* nothing, since we use xalloc.h */
@@ -129,7 +129,7 @@ format_check (void *msgid_descr, void *msgstr_descr, bool equality,
   struct spec *spec1 = (struct spec *) msgid_descr;
   struct spec *spec2 = (struct spec *) msgstr_descr;
   bool err = false;
-  unsigned int i;
+  size_t i;
 
   /* Check the argument types are the same.  */
   if (equality
@@ -148,7 +148,7 @@ format_check (void *msgid_descr, void *msgstr_descr, bool equality,
         {
           if (error_logger)
             error_logger (error_logger_data,
-                          _("format specifications in '%s' and '%s' for argument %u are not the same"),
+                          _("format specifications in '%s' and '%s' for argument %zu are not the same"),
                           pretty_msgid, pretty_msgstr, i + 1);
           err = true;
         }
@@ -191,9 +191,9 @@ get_sysdep_c_format_directives (const char *string, bool translated,
 
   if (descr != NULL && descr->sysdep_directives_count > 0)
     {
-      unsigned int n = descr->sysdep_directives_count;
+      size_t n = descr->sysdep_directives_count;
       struct interval *intervals = XNMALLOC (n, struct interval);
-      unsigned int i;
+      size_t i;
 
       for (i = 0; i < n; i++)
         {
@@ -227,7 +227,7 @@ static void
 format_print (void *descr)
 {
   struct spec *spec = (struct spec *) descr;
-  unsigned int i;
+  size_t i;
 
   if (spec == NULL)
     {
